made games editable & added modal for details on a game

+++ front/src/components/game/games-list.js
@@ -1,11 +1,21 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
+import { Redirect } from 'react-router-dom';
 import TableView from '../views/table-view';
+import ModalPopup from '../views/modal-view';
 
 
 export default function GamesList() {
 
   const [games, setGames] = useState([]);
+  const [redirect, setRedirect] = useState(null);
+  const [gameToEdit, setGameToEdit] = useState({});
+  const [modalShow, setModalShow] = useState(false);
+
+  const editGame = React.useCallback((game) => {
+    setGameToEdit(game)
+    setRedirect('/addGame')
+  }, []);
 
   const deleteOneGame = React.useCallback((gameId) => {
     axios.delete('http://localhost:5000/game/delete/', { data: { gameId: gameId } })
@@ -21,18 +31,20 @@ export default function GamesList() {
       return games.map(d => {
         return ({
           _id: d._id,
+          gameDetail: <button onClick={() => setModalShow(true)}>Details</button>,
           gameName: d.name,
           gameStatus: getAvailability(d),
           gameEditor: d.editor,
           gameAuthor: getAuthors(d),
           gameDescription: d.description,
+          gameEditButton: <button onClick={() => editGame(d)}>Modifier</button>,
           gameDeleteButton: <button onClick={() => deleteOneGame(d._id)}>Supprimer</button>
         })
       })
     } else {
       return []
     }
-  }, [games, deleteOneGame])
+  }, [games, editGame, deleteOneGame])
 
   const data = React.useMemo(() => processData(), [processData])
 
@@ -41,6 +53,10 @@ export default function GamesList() {
       {
         Header: '_id',
         accessor: '_id',
+      },
+      {
+        Header: '',
+        accessor: 'gameDetail',
       },
       {
         Header: 'Nom',
@@ -61,6 +77,10 @@ export default function GamesList() {
       {
         Header: 'Description',
         accessor: 'gameDescription',
+      },
+      {
+        Header: '',
+        accessor: 'gameEditButton',
       },
       {
         Header: '',
@@ -98,12 +118,14 @@ export default function GamesList() {
 
   function getAuthors(game) {
     let result = ''
-    game.authors.forEach(author => {
-      result === '' ? result = result + author : result = result + ' & ' + author
-    });
+    if (game.authors !== null) {
+      game.authors.forEach(author => {
+        result === '' ? result = result + author : result = result + ' & ' + author
+      });
+    }
     return result
   };
-  
+
   function getAllGames() {
     return new Promise(() => {
       axios.get('http://localhost:5000/games')
@@ -122,14 +144,23 @@ export default function GamesList() {
       })
   }
 
-  return (
-    <div>
-      <h2>Tous les jeux :</h2>
-      <br />
-      <TableView columns={columns} data={data} />
-      <br />
-      <button onClick={() => getAllGames()}>Rafraîchir</button>
-      <button onClick={() => deleteAllGames()}>Supprimer tous les jeux</button>
-    </div>
-  )
+  if (redirect) {
+    return <Redirect to={{ pathname: redirect, state: { referrer: gameToEdit } }} />
+  } else {
+    return (
+      <div>
+        <ModalPopup
+          show={modalShow}
+          onHide={() => setModalShow(false)}
+        />
+        <h2>Tous les jeux :</h2>
+        <br />
+        <TableView columns={columns} data={data} />
+        <br />
+        <button onClick={() => getAllGames()}>Rafraîchir</button>
+        <button onClick={() => deleteAllGames()}>Supprimer tous les jeux</button>
+        <button onClick={() => console.log(games)}>Console tous les jeux</button>
+      </div>
+    )
+  }
 };
